Cf/2025 05 21

+++ public/vocables/modules/tools/mappingModeler/tripleFactory.js
@@ -126,35 +126,35 @@ var TripleFactory = (function () {
                 alert(err.responseText || err);
             } else {
                 UI.message("Done", true);
-                if (!self.filterMappingIsSample) {
-                    //Admin.clearOntologyModelCache();
-                    SearchUtil.generateElasticIndex(MappingModeler.currentSLSsource, { indexProperties: 1, indexNamedIndividuals: 1 }, () => {
-                        $.ajax({
-                            type: "DELETE",
-                            url: `${Config.apiUrl}/ontologyModels?source=${MappingModeler.currentSLSsource}`,
+                var indexAuto = $("#MappingModeler_indexAutoCBX").prop("checked");
 
-                            dataType: "json",
-                            success: function (result, _textStatus, _jqXHR) {
-                                delete Config.ontologiesVocabularyModels[MappingModeler.currentSLSsource];
+                if (!self.filterMappingIsSample && indexAuto) {
+                    SearchUtil.generateElasticIndex(
+                        MappingModeler.currentSLSsource,
+                        {
+                            indexProperties: 1,
+                            indexNamedIndividuals: 1,
+                        },
+                        () => {
+                            $.ajax({
+                                type: "DELETE",
+                                url: `${Config.apiUrl}/ontologyModels?source=${MappingModeler.currentSLSsource}`,
 
-                                UI.message("ALL DONE");
-                            },
-                            error: function (err) {
-                                if (callback) {
-                                    return callback(err);
-                                }
-                                UI.message(err.responseText);
-                            },
-                        });
-                        /*
-                        $.ajax(`/api/v1/ontologyModels?source=${MappingModeler.currentSLSsource}`, { method: "DELETE" })
-                            .then((_success) => {
-                                window.UI.message(`${MappingModeler.currentSLSsource} was updated successfully`, true);
-                            })
-                            .catch((error) => {
-                                alert(error);
-                            });*/
-                    });
+                                dataType: "json",
+                                success: function (result, _textStatus, _jqXHR) {
+                                    delete Config.ontologiesVocabularyModels[MappingModeler.currentSLSsource];
+
+                                    UI.message("ALL DONE");
+                                },
+                                error: function (err) {
+                                    if (callback) {
+                                        return callback(err);
+                                    }
+                                    UI.message(err.responseText);
+                                },
+                            });
+                        },
+                    );
                 }
             }
         });
@@ -230,12 +230,16 @@ var TripleFactory = (function () {
                 MappingModeler.clearSourceClasses(DataSourceManager.currentSlsvSource);
                 if (callback) {
                     return callback();
+                } else {
+                    alert(result.result);
                 }
                 UI.message(result.result);
             },
             error: function (err) {
                 if (callback) {
                     return callback(err);
+                } else {
+                    alert(err.responseText);
                 }
                 UI.message(err.responseText);
             },
